feat: utilize ember-window-mock for the underlying provider
...of the services. This is _instead of_ custom implementations that
suited the needs of the maintainers of this package. ember-window-mock
uses a recursive proxy that allows any api to be stubbed, regardless of
its ability to be stubbed on the native object (ie: can't set a
read-only property).

fix: there is now URL validation on initial location.href values

BREAKING CHANGE: invalid hrefs will throw an error